+++ third/src/components/Task/Task.css
@@ -1,0 +1,3 @@
+.Task {
+    background-color: aquamarine;
+}
